requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
psycopg2==2.9.9
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1